refactor: :recycle: отделить локальные декларации от типов
Декларации перенесены к модулям-владельцам с сохранением прежних путей экспорта. Это устраняет четыре циклических компонента без изменения публичного поведения.
