+++ packages/core/metadata/appliedObjects/configuration/constants.ts
@@ -1,0 +1,1 @@
+export const CONFIGURATION_XML_FILE = "Configuration.xml"
