+++ packages/core/metadata/appliedObjects/metadataCatalog/rules.ts
@@ -21,7 +21,18 @@ import { V8_MDCLASSES_ROOT } from "../../orchestration/appliedObject/presets"
 import type { MetadataItemRule } from "../../orchestration/property/types"
 import { commonBasedOnObjectPaths } from "../../orchestration/metadataTarget"
 import { MetadataCommandRules } from "../metadataCommand/rules"
-import { MetadataCatalogStandardAttributeNames } from "./types"
+export const MetadataCatalogStandardAttributeNames: Record<string, string> = {
+  PredefinedDataName: "ИмяПредопределенныхДанных",
+  Predefined: "Предопределенный",
+  Ref: "Ссылка",
+  DeletionMark: "ПометкаУдаления",
+  IsFolder: "ЭтоГруппа",
+  Owner: "Владелец",
+  Parent: "Родитель",
+  Description: "Наименование",
+  Code: "Код",
+}
+
 export const MetadataCatalogRules = {
   itemType: "MetadataCatalog",
   metadataTargetOwner: { kind: "self", root: "Catalog" },
